requirements.txt
flask
gunicorn
requests
gspread
psycopg2-binary
python-dotenv
flask_login
